+++ mongo_db_restore.py
@@ -26,7 +26,8 @@
         -y value => A flavor id for the program lock.  To create unique lock.
         -v => Display version of this program.
         -h => Help and usage message.
-            NOTE 1:  -v or -h overrides the other options.
+
+        NOTE 1:  -v or -h overrides the other options.
 
     Notes:
         Warning:  If restoring to a Mongo database in a replica set, must
@@ -108,7 +109,6 @@ import subprocess
 # Local
 import lib.gen_libs as gen_libs
 import lib.gen_class as gen_class
-import lib.arg_parser as arg_parser
 import mongo_lib.mongo_class as mongo_class
 import mongo_lib.mongo_libs as mongo_libs
 import version
@@ -140,19 +140,18 @@ def single_db(server, args_array, **kwargs):
     Description:  Restore single database.
 
     Arguments:
-        (input) server -> Database server instance.
-        (input) args_array -> Array of command line options and values.
+        (input) server -> Database server instance
+        (input) args_array -> Array of command line options and values
         (input) **kwargs:
-            opt_arg -> Dictionary of additional options to add.
-            req_arg -> List of options to add to cmd line.
-        (output) False -> If an error has occurred.
-        (output) None -> Error message.
+            opt_arg -> Dictionary of additional options to add
+            req_arg -> List of options to add to cmd line
+        (output) False -> If an error has occurred
+        (output) None -> Error message
 
     """
 
     global AUTH_DB
 
-    subp = gen_libs.get_inst(subprocess)
     args_array = dict(args_array)
     req_arg = list(kwargs.get("req_arg", []))
     opt_arg = dict(kwargs.get("opt_arg", {}))
@@ -166,7 +165,7 @@ def single_db(server, args_array, **kwargs):
         arg_parser.arg_set_path(args_array, "-p"), req_arg=req_arg,
         opt_arg=opt_arg)
 
-    proc1 = subp.Popen(load_cmd)
+    proc1 = subprocess.Popen(load_cmd)
     proc1.wait()
 
     return False, None
@@ -179,11 +178,11 @@ def run_program(args_array, func_dict, **kwargs):
     Description:  Creates class instance(s) and controls flow of the program.
 
     Arguments:
-        (input) args_array -> Dict of command line options and values.
-        (input) func_dict -> Dictionary list of functions and options.
+        (input) args_array -> Dict of command line options and values
+        (input) func_dict -> Dictionary list of functions and options
         (input) **kwargs:
-            opt_arg -> Dictionary of additional options to add.
-            req_arg -> List of options to add to cmd line.
+            opt_arg -> Dictionary of additional options to add
+            req_arg -> List of options to add to cmd line
 
     """
 
@@ -216,12 +215,12 @@ def main():
         line arguments and values.
 
     Variables:
-        dir_chk_list -> contains options which will be directories.
-        func_dict -> dictionary list for the function calls or other options.
-        opt_arg_list -> contains optional arguments for the command line.
-        opt_req_list -> contains the options that are required for the program.
-        opt_val_list -> contains options which require values.
-        req_arg_list -> contains arguments to add to command line by default.
+        dir_perms_chk -> contains directories and their octal permissions
+        func_dict -> dictionary list for the function calls or other options
+        opt_arg_list -> contains optional arguments for the command line
+        opt_req_list -> contains the options that are required for the program
+        opt_val_list -> contains options which require values
+        req_arg_list -> contains arguments to add to command line by default
 
     Arguments:
         (input) argv -> Arguments from the command line.
@@ -230,31 +229,34 @@ def main():
 
     global AUTH_DB
 
-    cmdline = gen_libs.get_inst(sys)
-    dir_chk_list = ["-d", "-o", "-p"]
+    dir_perms_chk = {"-d": 5, "-o": 7, "-p": 5}
     func_dict = {"-S": single_db}
     opt_arg_list = {"-S": "--db=", "-o": "--dir="}
     opt_req_list = ["-c", "-d", "-o"]
     opt_val_list = ["-c", "-d", "-o", "-p", "-S", "-y"]
     req_arg_list = [AUTH_DB]
 
-    # Process argument list from command line.
-    args_array = arg_parser.arg_parse2(cmdline.argv, opt_val_list)
+    # Process argument list from command line
+    args = gen_class.ArgParser(sys.argv, opt_val=opt_val_list, do_parse=True)
 
-    if not gen_libs.help_func(args_array, __version__, help_message) \
-       and not arg_parser.arg_require(args_array, opt_req_list) \
-       and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list):
+    if not gen_libs.help_func(args, __version__, help_message)  \
+       and args.arg_require(opt_req=opt_req_list)               \
+       and args.arg_dir_chk(dir_perms_chk=dir_perms_chk):
+        
+#    if not gen_libs.help_func(args_array, __version__, help_message) \
+#       and not arg_parser.arg_require(args_array, opt_req_list) \
+#       and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list):
 
         try:
-            prog_lock = gen_class.ProgramLock(cmdline.argv,
-                                              args_array.get("-y", ""))
-            run_program(args_array, func_dict, opt_arg=opt_arg_list,
-                        req_arg=req_arg_list)
+            prog_lock = gen_class.ProgramLock(
+                sys.argv, args.get_val("-y", def_val=""))
+            run_program(
+                args, func_dict, opt_arg=opt_arg_list, req_arg=req_arg_list)
             del prog_lock
 
         except gen_class.SingleInstanceException:
             print("WARNING:  Lock in place for mongo_db_restore with id: %s"
-                  % (args_array.get("-y", "")))
+                  % (args.get_val("-y", def_val="")))
 
 
 if __name__ == "__main__":