+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [0.0.3] - 2023-10-19
 - Upgrade python-lib to v2.10.1
 - Upgrade mongo-libs to v4.2.7
+- Replaced the arg_parser code with gen_class.ArgParser code.
+
+### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
+- main, single_db: Removed gen_libs.get_inst call.
+- Documentation updates.
 
 
 ## [0.0.2] - 2021-12-22